requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
pandas==2.0.3
boto3==1.28.57
openpyxl==3.1.2
xlrd==2.0.1
